+++ src/Infrastructure/Bundle/InfrastructureBundle/Migrations/Version20200902093131.php
@@ -1,0 +1,26 @@
+<?php declare(strict_types=1);
+
+namespace Proximum\Vimeet\Infrastructure\Bundle\InfrastructureBundle\Migrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Tip: content can now be nullable
+ */
+final class Version20200902093131 extends AbstractMigration
+{
+    public function up(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE tip_translation CHANGE content content LONGTEXT DEFAULT NULL');
+    }
+
+    public function down(Schema $schema) : void
+    {
+        $this->abortIf($this->connection->getDatabasePlatform()->getName() !== 'mysql', 'Migration can only be executed safely on \'mysql\'.');
+
+        $this->addSql('ALTER TABLE tip_translation CHANGE content content LONGTEXT CHARACTER SET utf8 NOT NULL COLLATE `utf8_unicode_ci`');
+    }
+}