+++ src/Ui/Bundle/EventBundle/Form/Type/Participant/ParticipantTimezoneType.php
@@ -10,8 +10,8 @@
 
 namespace Proximum\Vimeet\Ui\Bundle\EventBundle\Form\Type\Participant;
 
-use Proximum\Vimeet\Application\Adapter\TranslatorInterface;
 use Proximum\Vimeet\Application\Command\Participant\ParticipantTimezone;
+use Proximum\Vimeet\Domain\Event\GetTimezoneHelper;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\TimezoneType;
 use Symfony\Component\Form\FormBuilderInterface;
@@ -19,19 +19,19 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class ParticipantTimezoneType extends AbstractType
 {
-    /** @var TranslatorInterface */
-    private $translator;
+    /** @var GetTimezoneHelper */
+    private $getTimezoneHelper;
 
-    public function __construct(TranslatorInterface $translator)
+    public function __construct(GetTimezoneHelper $getTimezoneHelper)
     {
-        $this->translator = $translator;
+        $this->getTimezoneHelper = $getTimezoneHelper;
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder->add('timezone', TimezoneType::class, [
-            'choice_label' => function ($key, $label) {
-                return $this->translator->trans('timezone.' . strtolower(str_replace('/', '-', $key)));
+            'choice_label' => function (string $key) {
+                return $this->getTimezoneHelper->getTimezoneTranslated($key);
             },
             'attr' => [
                 'class' => 'select2',