+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - VIMEET-1925 - Chat API
+- VIMEET-2106 - The speaking user video is maximized based on audio level
+- Deploy: DB migration without interaction
 
 ## [2.43.0] - 2020-08-27
 ### Added