+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VIMEET-2106 - The speaking user video is maximized based on audio level
 - Deploy: DB migration without interaction
 
+### Fixed
+- Tip content can be nullable, and tip translations affected on event come from the event locales.
+
 ## [2.43.0] - 2020-08-27
 ### Added
 - VIMEET-2187 - Invisible mode for webinar speaker