+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - VIMEET-2106 - The speaking user video is maximized based on audio level
 - Deploy: DB migration without interaction
 - VIMEET-2213 - Translate timezones
+- VIMEET 2178 - In a webinar, participant can vote for a question
 
 ### Fixed
 - Tip content can be nullable, and tip translations affected on event come from the event locales.